fix(STEP-3): Organisation config Spring Security
